+++ CHANGELOG.md
@@ -3716,3 +3716,4 @@
 - [2026-08-18 18:41] test(analytics): add unit tests for analytics business logic
 - [2026-08-18 18:42] feat(i18n): add async i18n stream provider
 - [2026-08-18 18:48] feat(navigation): integrate custom telemetry logger in navigation
+- [2026-08-18 19:19] feat(storage): integrate offline caching for storage state